SpringMVCExample/src/main/java/com/yash/controller/HelloWorldController.java
package com.yash.controller;

import java.util.ArrayList;
import java.util.List;
import java.util.Map;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.ModelMap;
import org.springframework.web.bind.annotation.ModelAttribute;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import com.yash.model.User;
import com.yash.service.UserService;

@Controller
public class HelloWorldController {
	@Autowired
	UserService userService;

	public void setUserService(UserService userService) {
		this.userService = userService;
	}

	@RequestMapping("/register")
	public String register(Map<String, Object> model) {
		User user = new User();
		model.put("User", user);

		List<String> professionList = new ArrayList<String>();
		professionList.add("Developer");
		professionList.add("Designer");
		professionList.add("IT Manager");

		model.put("professionList", professionList);
		return "register";
	}

	@RequestMapping(value = "/save", method = RequestMethod.POST)
	public String save(@ModelAttribute("User") User user, ModelMap model) {
		userService.save(user);
		model.put("username", user.getUsername());

		return "Save";

	}
}
